+++ src/ngraph/builder/reshape.cpp
@@ -24,10 +24,10 @@
 #include "ngraph/op/concat.hpp"
 #include "ngraph/op/constant.hpp"
 #include "ngraph/op/experimental/dyn_slice.hpp"
-#include "ngraph/op/experimental/shape_of.hpp"
-#include "ngraph/op/experimental/transpose.hpp"
 #include "ngraph/op/product.hpp"
 #include "ngraph/op/reshape.hpp"
+#include "ngraph/op/shape_of.hpp"
+#include "ngraph/op/transpose.hpp"
 #include "ngraph/opsets/opset1.hpp"
 #include "ngraph/util.hpp"
