+++ src/ngraph/op/ctc_greedy_decoder.hpp
