+++ src/ngraph/op/shape_of.hpp
