+++ src/ngraph/runtime/cpu/cpu_external_function.cpp
@@ -84,7 +84,6 @@
 #include "ngraph/op/experimental/quantized_conv_relu.hpp"
 #include "ngraph/op/experimental/quantized_dot_bias.hpp"
 #include "ngraph/op/experimental/random_uniform.hpp"
-#include "ngraph/op/experimental/tile.hpp"
 #include "ngraph/op/floor.hpp"
 #include "ngraph/op/fused/conv_fused.hpp"
 #include "ngraph/op/fused/gelu.hpp"
@@ -141,6 +140,7 @@
 #include "ngraph/op/sum.hpp"
 #include "ngraph/op/tan.hpp"
 #include "ngraph/op/tanh.hpp"
+#include "ngraph/op/tile.hpp"
 #include "ngraph/op/topk.hpp"
 #include "ngraph/op/xor.hpp"
 #include "ngraph/pass/algebraic_simplification.hpp"
