+++ src/ngraph/runtime/dynamic/dynamic_backend.cpp
@@ -23,9 +23,9 @@
 #include "ngraph/op/experimental/dyn_replace_slice.hpp"
 #include "ngraph/op/experimental/dyn_slice.hpp"
 #include "ngraph/op/experimental/generate_mask.hpp"
-#include "ngraph/op/experimental/range.hpp"
-#include "ngraph/op/experimental/transpose.hpp"
+#include "ngraph/op/range.hpp"
 #include "ngraph/op/reshape.hpp"
+#include "ngraph/op/transpose.hpp"
 #include "ngraph/pass/constant_folding.hpp"
 #include "ngraph/pass/dyn_elimination.hpp"
 #include "ngraph/pass/manager.hpp"
